Allow user to cancel during loads.
Also ignore Ctrl-C/SIGINT otherwise.

A keyboard interrupt will now just cause the current blocking operation
to cancel and return to the previous state.  For longer loads, or a
misbehaving network this is nicer then having no option then to restart
the application.

It also better fits expected behaviour as in that a text mode GUI
application doesn't quit on Ctrl-C.