pager: Add RTV_PAGER to support a reddit-specific setting
Reddit submissions and comments are styled with Markdown.  If you have a
Markdown-aware pager/viewer on your system it is convenient to be able to
use it rather than the default 'less'/'more'/etc. one used for everything
else on the system.  Support a high-priority environment variable that
will be consulted prior to checking $PAGER and/or defaulting to 'less'.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>