Implement a proper return key (fixes #689)
This PR implements a proper return key as proposed in issue #689.
As a consequence of that, the former (IMO rather complicated) approach
of using the methods `loop` and `handle_selected_page` is replaced
by using a simple page stack to keep track of the currently
selected page. This page stack also defines a maximal size: If
the page stack is full and an additional page is added to the page
stack, the **oldest** entry in the page stack will be removed, so
that the page stack does not exceed the maximal size.

For this new approach, the tests had to be adjusted. Here is a
summary:

Removed tests (no longer needed or replaced by the newly added tests):
 - test_inbox_close
 - test_submission_exit
 - test_subreddit_handle_selected_page
 - test_subreddit_page_loop_pre_select
 - test_subreddit_page_loop
 - test_subscription_close

Modified tests:
 - test_inbox_view_content
 - test_inbox_open_submission
 - test_submission_propmpt
 - test_subreddit_prompt
 - test_subreddit_open
 - test_subreddit_post
 - test_subreddit_open_subscriptions
 - test_subreddit_open_multireddits
 - test_subreddit_frontpage_toggle
 - test_subscription_prompt
 - test_subscription_close
 - test_page_unauthenticated

Added tests:
 - test_page_stack
 - test_page_back_button